Calibration Needed
Periodic Tasks has been redone to reflect the counter for moving up or
down a certain number of sensors. Teleop has been redone to support the
new region algorithm, which has been tested. The counter mentioned above
has not been implemented, but the arm can theoretically move to the
nearest sensors. Drive has also been edited in Teleop to allow for the
reversal of the drive train in the event of a pushing match. Vision is
currently not going to be included. The firing sequence has been
completed. Testing of the encoders still needs to happen, along with the
testing of the arm controls.